// src/_proto/Protos/notifications.ts
export interface GetMyNotificationsRequest {
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NotificationItem.Status status = 1;
     */
    status: NotificationItem_Status;
    /**
     * @generated from protobuf field: repeated string groups = 2;
     */
    groups: string[];
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.PageInfoRequest pageInfo = 3;
     */
    pageInfo?: PageInfoRequest;
}
// ...
export interface GetMyNotificationsCountRequest {
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NotificationItem.Status status = 1;
     */
    status: NotificationItem_Status;
    /**
     * @generated from protobuf field: repeated string groups = 2;
     */
    groups: string[];
}
// ...
export interface SetNotificationStatusesRequest {
    /**
     * @generated from protobuf field: google.protobuf.StringValue group = 1;
     */
    group?: StringValue;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NotificationItem.Status status = 2;
     */
    status: NotificationItem_Status;
}
// ...
export interface NotificationItem {
    /**
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * @generated from protobuf field: string objectId = 2;
     */
    objectId: string;
    /**
     * @generated from protobuf field: google.protobuf.Int32Value entityId = 3;
     */
    entityId?: Int32Value;
    /**
     * @generated from protobuf field: string type = 4;
     */
    type: string;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NotificationItem.Status status = 5;
     */
    status: NotificationItem_Status;
    /**
     * @generated from protobuf field: google.protobuf.StringValue actor = 6;
     */
    actor?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.StringValue slug = 7;
     */
    slug?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp createdOn = 8;
     */
    createdOn?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp timeRead = 9;
     */
    timeRead?: Timestamp;
    /**
     * @generated from protobuf field: map<string, string> data = 10;
     */
    data: {
        [key: string]: string;
    };
    /**
     * @generated from protobuf field: map<string, string> extraContent = 11;
     */
    extraContent: {
        [key: string]: string;
    };
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.NotificationItem.Status
 */
export enum NotificationItem_Status {
    /**
     * @generated from protobuf enum value: Unread = 0;
     */
    Unread = 0,
    /**
     * @generated from protobuf enum value: Read = 1;
     */
    Read = 1,
    /**
     * @generated from protobuf enum value: All = 2;
     */
    All = 2
}
// ...
class GetMyNotificationsRequest$Type extends MessageType<GetMyNotificationsRequest> {
    constructor() {
        super("wuxiaworld.api.v2.GetMyNotificationsRequest", [
            { no: 1, name: "status", kind: "enum", T: () => ["wuxiaworld.api.v2.NotificationItem.Status", NotificationItem_Status] },
            { no: 2, name: "groups", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "pageInfo", kind: "message", T: () => PageInfoRequest }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.GetMyNotificationsRequest
 */
export const GetMyNotificationsRequest = new GetMyNotificationsRequest$Type();
// ...
class GetMyNotificationsCountRequest$Type extends MessageType<GetMyNotificationsCountRequest> {
    constructor() {
        super("wuxiaworld.api.v2.GetMyNotificationsCountRequest", [
            { no: 1, name: "status", kind: "enum", T: () => ["wuxiaworld.api.v2.NotificationItem.Status", NotificationItem_Status] },
            { no: 2, name: "groups", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.GetMyNotificationsCountRequest
 */
export const GetMyNotificationsCountRequest = new GetMyNotificationsCountRequest$Type();
// ...
class SetNotificationStatusesRequest$Type extends MessageType<SetNotificationStatusesRequest> {
    constructor() {
        super("wuxiaworld.api.v2.SetNotificationStatusesRequest", [
            { no: 1, name: "group", kind: "message", T: () => StringValue },
            { no: 2, name: "status", kind: "enum", T: () => ["wuxiaworld.api.v2.NotificationItem.Status", NotificationItem_Status] }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.SetNotificationStatusesRequest
 */
export const SetNotificationStatusesRequest = new SetNotificationStatusesRequest$Type();
// ...
export const NotificationItem = new NotificationItem$Type();